# cars/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a simple CRUD API for profiles",
        "title": "Profiles API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:4200",
    "basePath": "/",
    "paths": {
...
        "/cars/add/{email}": {
            "post": {
                "description": "Post a user offer",
                "produces": [
                    "application/json"
                ],
                "summary": "Post an offer",
                "operationId": "post-offer",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Email address of the offer to be posted",
                        "name": "email",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "User data to be posted",
                        "name": "userData",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.Car"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/responses.UserResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/responses.UserResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/responses.UserResponse"
                        }
                    }
                }
            }
        },
...
        "/cars/details/{id}": {
            "get": {
                "description": "Get a user offer by id",
                "produces": [
                    "application/json"
                ],
                "summary": "Get an offer",
                "operationId": "get-offer",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Id of the offer to be retrieved",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/responses.UserResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/responses.UserResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/responses.UserResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/responses.UserResponse"
                        }
                    }
                }
            }
        },
...
        "/cars/search/user/{email}/{page}": {
            "get": {
                "description": "Get a user offers by email",
                "produces": [
                    "application/json"
                ],
                "summary": "Get offers by user",
                "operationId": "get-offers-by-user",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Email address of the offers to be retrieved",
                        "name": "email",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Page number",
                        "name": "page",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/responses.UserResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/responses.UserResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/responses.UserResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/responses.UserResponse"
                        }
                    }
                }
            }
        },
        "/cars/search/{page}": {
            "get": {
                "description": "Get offers by filter",
                "produces": [
                    "application/json"
                ],
                "summary": "Get offers",
                "operationId": "get-offers",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Page number",
                        "name": "page",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "name": "condition",
                        "in": "query"
                    },
                    {
                        "maximum": 1000,
                        "type": "integer",
                        "name": "distance",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "name": "doors",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "name": "drive",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "name": "engine_capacity_max",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "name": "engine_capacity_min",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "name": "filter_by",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "name": "fuel",
                        "in": "query"
                    },
                    {
                        "maxLength": 30,
                        "type": "string",
                        "name": "make",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "name": "mileage_max",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "name": "mileage_min",
                        "in": "query"
                    },
                    {
                        "maxLength": 30,
                        "type": "string",
                        "name": "model",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "name": "power_max",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "name": "power_min",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "name": "price_max",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "name": "price_min",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "name": "seats",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "name": "sort_direction",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "name": "steering",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "name": "transmission",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "name": "type",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "name": "year_max",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "name": "year_min",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/responses.UserResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/responses.UserResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/responses.UserResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/responses.UserResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "models.Car": {
            "type": "object",
            "required": [
                "description",
                "make",
                "model",
                "photos",
                "price",
                "title",
                "year"
            ],
            "properties": {
                "condition": {
                    "type": "string"
                },
                "description": {
                    "type": "string",
                    "maxLength": 3000,
                    "example": "Audi A4 2.0 TDI 2015, 190000 km, 10000 EUR"
                },
                "doors": {
                    "type": "integer"
                },
                "drive": {
                    "type": "string"
                },
                "engine_capacity": {
                    "type": "integer"
                },
                "first_registration": {
                    "type": "string"
                },
                "fuel": {
                    "type": "string"
                },
                "location": {
                    "$ref": "#/definitions/models.Location"
                },
                "make": {
                    "type": "string",
                    "maxLength": 30,
                    "example": "Audi"
                },
                "mileage": {
                    "type": "integer"
                },
                "model": {
                    "type": "string",
                    "maxLength": 30,
                    "example": "A4"
                },
                "photos": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    },
                    "example": [
                        "https://example.com/image.jpg"
                    ]
                },
                "power": {
                    "type": "integer"
                },
                "price": {
                    "type": "integer",
                    "example": 10000
                },
                "registration_number": {
                    "type": "string"
                },
                "seats": {
                    "type": "integer"
                },
                "steering": {
                    "type": "string"
                },
                "telephone_number": {
                    "type": "string"
                },
                "title": {
                    "type": "string",
                    "maxLength": 40,
                    "example": "Audi A4 2.0 TDI"
                },
                "transmission": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                },
                "vin_number": {
                    "type": "string"
                },
                "year": {
                    "type": "integer",
                    "example": 2015
                }
            }
        },
...
        "models.Location": {
            "type": "object",
            "properties": {
                "coordinates": {
                    "type": "array",
                    "items": {
                        "type": "number"
                    },
                    "example": [
                        0
                    ]
                },
                "type": {
                    "type": "string",
                    "example": "Point"
                }
            }
        },
        "responses.UserResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "object",
                    "additionalProperties": true
                },
                "message": {
                    "type": "string"
                },
                "status": {
                    "type": "integer"
                }
            }
        }
    }
}